Updated classes to stop using MDItem, MDFile, etc. and use HLKit instead.
Moved the handling of the item icons from MDHLDocument to HKItem.

Updated copyright year and project name in HLKit files.

Changed the properties of HKItem to nonatomic to improve performance.

Added an HK_LAZY_INIT compiler option for HKFile and HKFolder. With it enabled, the properties like name, nameExtension, kind, fileType, size, and type aren't set until they are asked for. This should help performance dramatically.

Added the higher-level file and folder additions to HKFileAdditions such as sound, image, and movie.

 

git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.0.3@68 c3d3028b-b5cd-4c35-8691-6fb8b2347b27