Removed references to capability to disable Exposé, Spaces, and Dashboard, and Using Sound Workaround and Preventing Gamma Correction which were never implemented and likely never will be.
Added a @synchronized() directive around the +defaultManager class method for thread safety.

Updated the helpGame::: method when helping for USB Overdrive to work with the new bundled proxy app I created that better explains when a user double-clicks on it in the Finder that it isn't intended to be used that way.



git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.0.3@172 c3d3028b-b5cd-4c35-8691-6fb8b2347b27