Updated copyright.
Changed properties to nonatomic and made them readonly to clarify that they should not be changed directly.

Removed the comparison method that took options.

Added a TKDXTCompressionQualityFromString() function to translate between string representation and TKDXTCompressionQuality enum type.

Revised the TKPixelFormat enumerated types; these still need work.

Added the following 2 class methods to allow creation of TKImageReps from existing NSBitmapImageReps:

/* create TKImageRep(s) from NSBitmapImageRep(s) */
+ (id)imageRepWithImageRep:(NSBitmapImageRep *)aBitmapImageRep;
+ (NSArray *)imageRepsWithImageReps:(NSArray *)bitmapImageReps;

Added a method to generate an array of mipmaps image reps of the receiver.

Added a preliminary method to generate an array of image reps by applying a normal map filter to them. There are potentially 2 libraries to choose from: Nvidia texture tools, and my own algorithm, perhaps written in CoreImage or using the Accelerate library. This is still very much in the works.

Added a convenience method to find the image rep that represents a particular face in a cube map from a given array of image reps.

.m:

representationUsingPixelFormat: still needs work.

Fixed a bug in isEqual: that affected Leopard systems and earlier.


git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.5@222 c3d3028b-b5cd-4c35-8691-6fb8b2347b27