Updated copyright.
Changed properties to nonatomic and made them readonly to clarify that they should not be changed directly.

An attempt to add additional DDS formats enumerated types is still in the works.

Added an options: NSDictionary parameter to the DDSRepresentationsOfImageRepsInArray:: method.

TKDDSImageRep is the class which actually implements the following mipmap generation method that TKImageRep declares:

- (NSArray *)mipmapImageRepsUsingFilter:(TKMipmapGenerationType)filterType;

The TKImageRep instance creates a TKDDSImageRep from itself and "forwards" the method call to the newly created dds image rep.

NOTE: both the mipmap and normal map generation features are still very early in development and may not work correctly yet.

.m:

Changed as many symbol names and functions to be static as possible.

Added @synchronized() directive around defaultFormat and setDefaultFormat: class methods for thread safety.


git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.5@225 c3d3028b-b5cd-4c35-8691-6fb8b2347b27