Updated copyright.
TKImageDocument:

Updated class to no longer use the <TKImageViewAnimatedImageDataSource> protocol. It now uses the new `animationImageReps` property of the TKImageView.

Added support for utilizing the TKImageInspectorController class by conforming to the <TKImageInspectorDataSource> protocol.

Added a third “faces” IKImageBrowserView on the left side of the window.

Added blue background views similar to how Preview implements its image-browser type interface.

Added a `normalMapPanel` and `grayscalePanel` panels which are HUD-style and allow application of filters to the images. Still in early development.

Added instance variables and properties for the various parameters used in the above-mentioned filter panels.

To optimize performance and make life easier, we know have 2 mutable arrays, `visibleFaceBrowserItems` and `visibleFrameBrowserItems`. When the user selects different faces and then different frames, these mutable arrays are updated to reflect their selection. That allows easy filtering of the items returned for the data source methods.

Added the methods of the informal TKImageInspectorController protocol/category.

Removed unused IBAction methods like `sendImageTo:`, `openInNewWindow:`, etc., which were modeled after Preview’s contextual menu items but which were never implemented.

Changed to use the TKAppKitAdditions.h rather than the MDAppKitAdditions.h.

Added a newly designed TKImageDocumentLion nib file which enhancements for OS X Lion and greater.

Factored creation of blue background layer into function MDBlueBackgroundLayerWithFrame().

Added a workaround for the “This feature is a work in progress” alert that is shown when the window controller finishes loading the nib file. In later versions of OS X, the alert ended up floating in mid air not attached to a window properly, so I added a performSelector:afterDelay: call which seems to have fixed it.

Added an `importImageSequence:` method to import a sequence of images to use in an animation sequence.

Began to add <NSSplitviewDelgate> methods to attempt to better manage the split views. Still in progress, I believe.

Started to implement the following <IKImageBrowserDataSource> methods

imageBrowser:removeItemsAtIndexes:
imageBrowser:moveItemsAtIndexes:toIndex:

Added support to keep track of the currently selected browser view (as the `currentMenuBrowserView` ivar).

Added support for dragging image files from the Finder into the browser views in the image document window. Loading of the images is done in a background thread.

Added support for generating and removing mipmaps via menu item.

All add, remove, and move operations are isolated in the later part of the class implementation. I hope for most operations to be undoable via the undoManager.

Implemented validateMenuItem: method.



git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.5@236 c3d3028b-b5cd-4c35-8691-6fb8b2347b27