Added support for an “Original Image” preset, which represents the original image without any compression. To support this, IBOutlets for NSTextFields were added to be able to hide them.
Added a mipmaps style popup button.

Changed to use “TKAppKitAdditions.h” in place of “MDAppKitAdditions.h”.

Added a `mediator` NSObjectController and a `zoomMediator` NSObjectController.

Added a `previewViewZoomFactor` ivar to attempt to replicate Photoshop’s Save For Web feature, where the zoom level stays consistent across all 2 or 4 preview panes. Not sure if it’s working yet fully or not.

Changed the declared property for the `preset` to be retain rather than copy.

Added a changeTool: IBAction for the tools. Not yet implemented.

Added a cleanup method to handle the unbinding, and removal of notification observers, in an attempt to eradicate the crashing bug.

When the 2 or 4 TKImageExportPreviewViewController are allocated, the imageView’s zoomFactor is bound to the zoomMediator NSObjectController.

Added code to set the nextKeyView sequence of the preview views.

Updated the synchronizeUI method to better support the Original Image preset.

Calling the setPreset: method now automatically calls [self synchronizeUI] to keep the UI synchronized.

Removed the `previews` ivar and instead deal with the `representedObject` of the corresponding previewController in the `previewControllers` array.

Added a consolidated `saveDefaults` method which now saves the arrangement of presets of the 2 or 4 preview controllers and the frame size of the sheet view.


git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.5@240 c3d3028b-b5cd-4c35-8691-6fb8b2347b27