Removed the TKImageViewAnimatedImageDataSource protocol and simply created a `animationImageReps` property similar to how UIImageView does it in iOS.
Added a `showsImageBackground` feature which is similar to Preview. It shows a checkerboard pattern behind the transparent image (or is supposed to anyway).

I’ve realized that I technically hacked the private details of the IKImageView layer implementation on OS X 10.6 or whichever SDK this was originally developed with. Now that I have linked with a newer SDK, the implementation details have changed and the animation feature no longer works properly. This will need some work.


git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.0.3@261 c3d3028b-b5cd-4c35-8691-6fb8b2347b27