Updated to HLKit 2.4.4.
Removed HKFoundationAdditions.h and HKPrivateInterfaces.h from the private header files and moved to project header files (non-distributed).

Revised imports on other source files that used HKFoundationAdditions.h.

Added a "md_" prefix to NSString (HKAdditions) category methods to try to avoid any conflicts in the future.

Fixed the conflicting signatures on the private HKFile and HKFolder `init` methods: they were defined to take a `const CDirectoryFolder *`, and I had them implemented without the `const`.

Added a (__bridge void *) cast to the assigning of `_privateData` (not sure why this was necessary, as I'm not using ARC, but anyway).

Conflicts:
	Source Finagler.xcodeproj/project.pbxproj