Updated SteamKit.framework to work with new locations of Valve's games.
Updated VSGame to use all URLs instead of NSString-based paths.

Moved VSGame*Key defines from VSSteamManager.h to VSGame.h, and changed as many of them as possible to be static rather than exported.

Made VSGame's and VSSteamManager's ivars @private.

Added an `appManifestURL` property to VSGame to keep track of the presence of an `app manifest_<gameID>.acf` app manifest file in the root of the SteamApps folder. If this property is non-nil, it means the VSGame has an upgraded location.

Added an appManifestURL: parameter to the VSGame init methods.

Added a VSGameLongName key/value pair to the entries in the com.valvesoftware.games.plist file. This is the long name that is used in the folder name of the game in its upgraded location.

Added a VSGameIconName key/value pair to the entries in the com.valvesoftware.games.plist file. This specifies the name of the icon file that should be used to represent the game in the list and to potentially be added to the resource fork of the helped executableURL. (Generally, the icon file name is simply `game.icns`, but Garry's Mod is different). Previously, I had hard-coded the game icon filename as a constant in VSSteamManager.

Updated VSSteamManager to work with the new "upgraded" location and installation method of Valve's games.

Added some more comments to make code clearer to others.

Updated VSSteamManager to work with URL-based VSGames.