Finished updating the feature that allows TKImageView to display animated VTF images.
In its original implementation, I was attempting to manipulate the private CALayer that TKImageView inherits from IKImageView, in order to display the animated images inside a core animation in this layer. While this worked fine in an earlier version of OS X, it broke in recent versions. As such, the new approach to accomplishing the animation feature treats the IKImageView’s CALayer as private and does not attempt to mess with it. Instead, TKImageView adds a subview of class TKAnimationImageView when it needs to display the animated images, and removes it when it stops displaying the animated images.