SteamKit: Updated framework classes to use enhanced `HKVPKFile` class and `VSSourceAddon` class.
Added the VSSourceAddon.h #import to the SteamKit.h header file as well as adding the _VSSourceAddon symbol name to SteamKit.exp exported symbol file.

Renamed `VSGame`’s `addonsFolderURL` to the more appropriate `sourceAddonsFolderURL`.

Changed the `VSGame`’s properties that are implicitly `assign` to be explicitly `assign` for clarification.

Changed the `VSSourceAddon` class to be URL-based similar to `VSGame` (`VSSteamManager` isn’t quite there yet).

For installing Source Addon (.vpk) files, the `VSSteamManager` now has a new method `installSourceAddon:usingMethod:` that takes a `VSSourceAddon` file and installation method parameter. Installation (copying of file contents) is now done in a background thread, and 2 new `<VSSteamManager>` delegate methods are used to inform the client of installation status asynchronously. Most of the examination and validation work that was previously done by the `VSSteamManager` class to assure that the .vpk was valid and to get the sourceAddonGameID, etc. has been moved to the enhanced `HKVPKFile` and `VSSourceAddon` classes.

Previously, the `VSSourceAddon` class was a fairly dumb class that simply served as placeholders for the values that would be shown in Source Addon Finagler’s table view: for example, `problem` text, `fileIcon`, etc. The class has now been re-written to fit in closer with the `VSGame` class and to use the new enhanced `HKVPKFile` class to discover its properties. Value Transformers will be used in the updated interface where necessary to produce human-readable results.

Updated the `NSString` error description text that is returned in `VSSteamManager`’s `isProposedRelocationPathValid:errorDescription:` to be localized, which also means I added a Localizable.strings file to the project.

Removed commented-out code and unused constants.

Enhanced the `description` method with new properties being printed.