Source Addon Finagler: Updated target to work with new VPK enhancements.
Added the 2 styles of schemes to Source Addon Finagler like I did for Source Finagler: Debug and Release.

Created a Localizable.strings file that is specific to the Source Addon Finagler app, rather than just using the same one as Source Finagler itself. This file is stored in the `en.lproj` subdirectory of the `SourceAddonFinagler` folder.

Moved the `VSSourceAddonFinaglerMainMenu.xib` file from the root `English.lproj` folder to the `en.lproj` subfolder in the `SourceAddonFinagler` folder.

Updated the `VSSourceAddonController` class to work with the new capabilities of `VSSteamManager` and `VSSourceAddon`: `VSSourceAddonController` now conforms to the `<VSSteamManager>` delegate protocol to implement the callback methods for progress of installing addons. The main progress window now shows a determinate progress bar if multiple addons are being installed.

I simplified the interface configuration and layout by always starting out with an initial 3-section split view and removing the subviews depending on the number of different installation outcomes: successfully installed, already installed, and not installed. As subviews are removed, the height of the window is also reduced accordingly. Once the proper number of subviews is achieved, the boxes are filled with the appropriate views.

Added a third table view to show Source Addons that were already installed. I realized that previously, if a user were to double-click on a Source Addon that was already “installed” (inside the appropriate `addons` folder), I would consider that an error, saying something like “Couldn’t install the addon because it’s already installed”, which is kind of an inelegant way of handling it. So, I added the ability to differentiate the attempt to install addons that are already installed.

I switched from using the data source tableview methods to using bindings and array controllers.

Added a new `VSSourceAddonTransformer`, that handles generating the problem text for Source Addons that weren’t installed. It takes in a `VSSourceAddon` object and generates the string depending on the `VSSourceAddon`’s `sourceAddonStatus`.

Previously, Source Addon Finagler had the highest priority and would be the default app for opening VPK files, which, now that we know that multi-segment VPK archives are the default format from here on out, is undesirable. I’ve changed Source Addon Finagler to be a Viewer of VPK files, and changed its rank from Owner to Alternate. Source Finagler is now the Owner and default app for VPK files.

Added an "Install Source Addon" service to the Source Addon Finagler application. This will allow a user to right-click on a Source Addon file in the Finder and choose to install it directly, rather than having to use a 2-step process of opening it first in Source Finagler and then clicking the Install Source Addon button.