HLKit: Updated `HKArchiveFile` and its concrete subclasses.
In `HKArchiveFile`, renamed the `fileType` property to the more appropriate `archiveFileType`.

Changed the `+fileTypeForData:` method to `+archiveFileTypeForContentsOfFile:` which takes the file path to the file in question rather than just raw data. This is necessary because the “*_001.vpk” segments of multi-part VPK archives can begin with any kind of data, not necessarily a VPK magic header data.

Changed the `filePath` and `archiveFileType` properties to be nonatomic, and the `version` property to be readonly and nonatomic.

Updated the `description` method in concrete subclasses to at least add in super’s description to theirs.