Source Finagler: Updated target to work with new VPK enhancements.
Changed the `CFBundleTypeName` (name) of the .vpk file from `Source Addon file` to the more appropriate `Source Valve Package file`. It turns out my original understanding of the VPK archive format as being only for Source Addons was incorrect; the VPK format is the new standard distribution format that solves the problems present in both of the older archive formats, GCF and NCF.

Since VPK is the de-facto standard archive format from here on out, I’ve enhanced the `MDVPKDocument` class to better handle situations where the .vpk file opened is part of multi-segment VPK archive as opposed to being a self-contained Source Addon file. Multi-segment VPK archives are composed of a “table of contents” (directory) archive whose filename ends with `_dir.vpk`, and which references the multiple segments which will have a suffix like `_nnn.vpk`, where n is the segment number like `001` or `014`. With a multi-segment VPK archive, you need to open only the “archive directory” file (the one that ends with `_dir.vpk`), not any of the multi-segment files, which when opened would appear to have empty contents. To help clarify to the user that this is the case, I’ve added code to the `MDVPKDocument` class to override setup and teardown in cases where the VPK document is multipart archive and can't be viewed. If the user opens a multi-segment that is not the archive directory, `MDVPKDocument` creates an instance of `MDVPKViewController` which provides an interface explaining to the user that they should open the `*_dir.vpk` file instead. It provides an option to automatically always do this in the future, if desired.

Updated the comments in some of the files to remove references to classes and objects from my Font Finagler project.

Moved the .nib files and localized resource files from the root `English.lproj` folder to the target-specific subfolder (`SourceFinagler/en.lproj`).

Regenerated the `Localizable.strings` file from scratch, which has rearranged the order of many of the string entries. This will be the new order going forward, since that’s the way `genstrings` will generate them, and there’s no sense in rearranging them to match what I originally had.

Renamed some of the internal Interface Builder names of objects in the nib files for the bottom bar and browser.

Updated the `MDBottomBar` class to handle the case where the passed in `totalCount` `NSNumber` value is `nil`, which is the case when the user mistakenly opens one of the segments of a multi-segment VPK archive. It now simply displays the amount of available disk space without attempting to show the number of items.

Cleaned up many of the classes by removing unused code that was residual from my Font Finagler project.

Previously, Source Addon Finagler had the highest priority and would be the default app for opening VPK files, which, now that we know that multi-segment VPK archives are the default format from here on out, is undesirable. I’ve changed Source Addon Finagler to be a Viewer of VPK files, and changed its rank from Owner to Alternate. Source Finagler is now the Owner and default app for VPK files.

Removed logging code from `MDStatusImageView`.

Removed code in `MDStatusImageView` that dealt with tracking areas, as it looks like I experimented around with that but left it in.

Moved initialization of `NSUserDefault` values dealing with list view font size and icon size, and browser font and icon size, from the main app controller to the `MDBrowser` and `MDOutlineView` classes themselves.

Removed all use of an undo manager in the `MDHLDocument` class.

Added an additional status image view to show when a file represents a Source Addon file. This status view is hidden by default, but `MDVPKDocument` will show it when applicable. Added a preliminary `sourceAddonIndicator.png` image file.

Added a button to the bottom bar to allow a user to install the Source Addon file using the Source Addon Finagler app. This button is hidden by default, but  `MDVPKDocument` will show it when the file represents an uninstalled Source Addon file.

Reworked `MDVPKDocument`’s `windowControllerDidLoadNib:` method to set an appropriate minimum size for the document window. Previously, you were able to resize the window to make it small enough that it would distort the alert view that was shown in the box.

Added preliminary code in an attempt to workaround the bug with `NSOutlineView` restoring autosaved columns in the incorrect order. While I’m still investigating this bug, the code will now make sure that at least the Name column is the first table column.

Added a delay to the setup code in `MDHLDocument` that prevents the reordering of the first table column. This allows the outline view to make sure the Name column is the first table column by reordering them if necessary upon load.

Updated `MDHLDocument`’s `validateMenuItem:` method to disable the newer “versioned documents” menu item commands that used selectors such as `duplicateDocument:`, `renameDocument:`, etc. Choosing these prior to this would cause an exception because file saving is not implemented (nor do I intend it to be).