MDInspectorView: Refactored to allow for better subclassing.
Updated the `MDInspectorView` to allow for better subclassing: I changed to not using any setters or getters in `init` methods, but instead, accessing ivars directly. In all other methods, on the other hand, I now use the getter methods, to allow subclasses to easily override methods (namely the `autosaveName` method).