TKImageExportPanel.xib and TKImageExportPreviewView.xib: Removed dependency on my MDAppKit.framework and MDAppKitIBPlugin.ibplugin.
Originally, I wrote a custom Interface Builder 3.x plugin to allow me to set up and encode instances of my `MDInspectorView`, `MDDateFormatter`, and `MDFileSizeFormatter` classes into nib files. While this was convenient for me, other developers won't be able to open the nib files without having my MDAppKit.framework present on their system. Since all 3 of the aforementioned classes are included in this Source Finagler project, I decided to re-implement the nib files to eliminate this dependency.

In the original `TKImageExportPanel.xib` file, I had a single instance of a preview view (which is now contained within the `TKImageExportPreviewView.xib` file) to use as reference. I removed this unused view to remove dependency on plugin.

In the original `TKImageExportPreviewView.xib` file, I had encoded a custom instance of `MDFileSizeFormatter`. I removed it from the nib file, added an `IBOutlet` to the `imageFileSizeField`, and now set the formatter programmatically in `awakeFromNib`.