BUGFIX: Fixed the conditional define in `MDFileManager` which determined which `NSFileManager` methods to call.
In the `MDFileManager` class, I wanted to use the new functionality and methods available in OS X 10.5 and later whenever possible, and if I was targeting an earlier version of OS X like 10.4, I would resort to using the old methods. Originally, to try to achieve this, I had been using the following define around my code:

#if MAC_OS_X_VERSION_10_5 <= MAC_OS_X_VERSION_MAX_ALLOWED
	fileManager = [[NSFileManager alloc] init];
#else
	fileManager = [[NSFileManager defaultManager] retain];
#endif

However, that's an incorrect way of going about it. What that equates to is basically trying to assure that the SDK I'm building against is OS X 10.5 or greater, and disregarding the important thing, which is the minimum deployment target version of OS X I want to still support. It instead should be written like the following:

#if MAC_OS_X_VERSION_MIN_REQUIRED >= 1050
	fileManager = [[NSFileManager alloc] init];
#else
	fileManager = [[NSFileManager defaultManager] retain];
#endif