SourceSpotlightImporter: Updated the logging methods.
Updated the logging methods to precede the log message with the "com.markdouma.mdimporter.Source " bundle identifier to make it easier to identify which importer is printing these messages.

Updated the logging method to include a dynamic method of determining the function name (using __FUNCTION__) rather than hard-coding it in the code.