Made Xcode schemes (.xcscheme) shared rather than user-specific. See notes below.
NOTE: The deployment target for Source Finagler 2.0.3 is 10.5, for both i386 and x86_64. To produce an app that properly runs in 10.5, my testing shows it needs to be compiled with gcc using Xcode 3.2.6. That said, I tend to do most of my editing of code in Xcode 5, hence the schemes added here.