SourceSpotlightImporter: updated to work with new capabilities of TextureKit.
Added a new `com_markdouma_image_dds_dx_ten_header` boolean value for whether a DDS file contains a DirectX 10 header.

Updated the `TKSFTextureImageType` section to call the new methods of TKImage that include the error parameter.

Added a value for `kMDItemPixelCount` to the attributes, but only in TKSnowLeopard and later.