HLKit: updated `HKArchiveFile` and its concrete subclasses.
In `HKArchiveFile`, renamed the `fileType` property to the more appropriate `archiveFileType`.

Changed the `+fileTypeForData:` method to `+archiveFileTypeForContentsOfFile:` which takes the file path to the file in question rather than just raw data. This is necessary because the “*_001.vpk” segments of multi-part VPK archives can begin with any kind of data, not necessarily a VPK magic header data.

Changed the `filePath` and `archiveFileType` properties to be nonatomic, and the `version` property to be readonly and nonatomic.

Updated the `description` method in concrete subclasses to at least add in super’s description to theirs.

Added a `fileSize` property to the `HKArchiveFile` class to represent the overall file size of the archive. This is intended to be used by the improved `VSSourceAddon` class and ultimately by the Source Addon Finagler app to help provide progress of copying or moving files based on file sizes.

Improved `HKFile`'s `stringByExtractingToTempFile:` method to use the filename of the container archive file in the name of the temp file, and in the error messages that are logged to console. Provided this method does fail to determine the string encoding of the specified `HKFile`, the extra filename information should help specify the actual archive file and file that is experiencing the problem.