HLKit: BUGFIX: fixed bug that could cause items being extracted to be written improperly or not at all.
Switching to a "lazy init" approach, where the `name` of an item was only loaded upon first request, resulted in the `pathRelativeToItem:` possibly returning `nil` rather than the actual name. The end result of this is that if a user hadn't fully browsed the hierarchy to the items they wanted to extract, an item with the name of folder would be written to the filesystem, yet would actually contain the contents of one of the items to be extracted. This was a rather odd bug that would sometimes show up and sometimes not. I've fixed it by making sure to use the accessor via `self.name` rather than trying to access the `name` instance variable directly.