Source Addon Finagler: updated target to work with new VPK enhancements.
Created a `Localizable.strings` file that is specific to the Source Addon Finagler app, rather than just using the same one as Source Finagler itself. This file is stored in the `en.lproj` subdirectory of the `SourceAddonFinagler` folder.

Moved the `VSSourceAddonFinaglerMainMenu.xib` file from the root `English.lproj` folder to the `en.lproj` subfolder in the `SourceAddonFinagler` folder.

Updated the `VSSourceAddonController` class to work with the new capabilities of `VSSteamManager` and `VSSourceAddon`: `VSSourceAddonController` now conforms to the `<VSSteamManager>` delegate protocol to implement the callback methods for progress of installing addons. The main progress window now shows a determinate progress bar if multiple addons are being installed.

Added a third table view to show Source Addons that were already installed. I realized that previously, if a user were to double-click on a Source Addon that was already “installed” (inside the appropriate `addons` folder), I would consider that an error, saying something like “Couldn’t install the addon because it’s already installed”, which is kind of an inelegant way of handling it. So, I added the ability to differentiate the attempt to install addons that are already installed.

I simplified the interface configuration and layout by revising the `finishInstallation` method. I removed the use of `NSBox`s and just keep `IBOutlet`s to the `splitView`, and the top-level views: `installedView`, `alreadyInstalledView`, and `problemView`. I check the count of each type of installed addon result and if present, the corresponding view gets added to the subviews to set the `NSSplitView` to using `setSubviews:`. After setting the subviews, I call `[splitView adjustSubviews]` to adjust the split view accordingly. If the window is not currently visible, I also reduce the height of the window according to the number of subviews I will set the split view to have.

I switched from using the data source tableview methods to using bindings and array controllers.

Added a new `VSSourceAddonTransformer`, that handles generating the problem text for Source Addons that weren’t installed. It takes in a `VSSourceAddon` object and generates the string depending on the `VSSourceAddon`’s `sourceAddonStatus`.

Previously, Source Addon Finagler had the highest priority and would be the default app for opening VPK files, which, now that we know that multi-segment VPK archives are the default format from here on out, is undesirable. I’ve changed Source Addon Finagler to be a Viewer of VPK files, and changed its rank from Owner to Alternate. Source Finagler is now the Owner and default app for VPK files.

Added an "Install Source Addon" service to the Source Addon Finagler application. This will allow a user to right-click on a Source Addon file in the Finder and choose to install it directly, rather than having to use a 2-step process of opening it first in Source Finagler and then clicking the Install Source Addon button. (NOTE: this only appears to be available on OS X 10.6+).

Added the setting and saving of the table view sort descriptors.