PROJECT: massive internal changes to better prepare the 2.5 branch for having the 2.0.3 release branch merged into it.
Went through the project.pbxproj files side by side and changed the internal identifiers (like `C1097AD6125FA15700D60D27`) to be the same in both projects for the corresponding files, groups, targets, file references, dependencies, build files, frameworks, container item proxies, etc. This should have no effect on the actual built products, which should still be identical; rather, these changes are only to bring the parts of the projects that are intended to do the same thing to actually be the same, which will make for much easier merging in the future.

Moved MDCGImage files from root level of the project folder to the SourceImageQuickLook folder.

Removed the Syntax Definitions folder.

Removed extraneous images from the Resources folder.

Removed the unused `ddsFinagler` target and scheme files from the project.