PROJECT: massive internal changes to better prepare the 2.0.3 branch for having the 2.5 branch merged into it.
Went through the project.pbxproj files side by side and changed the internal identifiers (like `C1097AD6125FA15700D60D27`) to be the same in both projects for the corresponding files, groups, targets, file references, dependencies, build files, frameworks, container item proxies, etc. This should have no effect on the actual built products, which should still be identical; rather, these changes are only to bring the parts of the projects that are intended to do the same thing to actually be the same, which will make for much easier merging in the future.

Removed the config.icns icon which is no longer used.