HLKit: BUGFIX: fixed bug that could cause crash when opening some multipart archives (i.e. "blah_001.vpk" rather than "blah_dir.vpk").
Fixed a crash that would occur when the outline view of an MDVPKDocument would ask for the items in the datasource methods, which would call down into the CVPKFile instance, which would be invalid in some cases if the .vpk file in question was a segment from a multipart .vpk file (for example, the "blah_001.vpk" segment instead of the "blah_dir.vpk" directory file). In the case of opening an HKVPKMultipartArchiveType type .vpk file, we no longer create a CVPKFile for it.

Improved the `description` method to append descriptions of the `vpkArchiveType` and `sourceAddonStatus` properties.