Source Finagler: preliminary tweaks to MDHLDocument.
Changed the document view mode to an actual `MDHLDocumentViewMode` enumerated type and added `MDHLDocumentNoViewMode` to signify no view mode (which is what a part of a multi-segment VPK file will use).

Made the `viewMode` and `selectedItems` to be readonly properties of the MDHLDocument class.

Changed other properties to be readonly if applicable.

Removed properties that don't need to be public (or properties at all).

Created a `reloadOutlineViewData` method that is called that handles setting the `outlineViewIsReloadingData` variable automatically.

Added a description of the `viewMode` property to the description method.