Source Finagler: removed global vars in MDAppController for MDHLDocument-related settings (show view options, inspector, etc.)
I got rid of almost all of the global variables in MDAppController that dealt with MDHLDocument-related settings such as MDShouldShowViewOptions, MDShouldShowInspector, MDShouldShowQuickLook, MDShouldShowPathBar, and MDPlaySoundEffects. Instead of being global variables stored in MDAppController, they are now static variables that can, in a way, be thought of as belonging to the MDHLDocument class itself. I added the following set of methods to MDHLDocument to allow access to these variables:

	+ (BOOL)shouldShowViewOptions;
	+ (void)setShouldShowViewOptions:(BOOL)shouldShow;

	+ (BOOL)shouldShowInspector;
	+ (void)setShouldShowInspector:(BOOL)shouldShow;

	+ (BOOL)shouldShowQuickLook;
	+ (void)setShouldShowQuickLook:(BOOL)shouldShow;

	+ (BOOL)shouldShowPathBar;
	+ (void)setShouldShowPathBar:(BOOL)shouldShow;

I use NSRecursiveLocks to protect access to these variables, unlike the previous accessing of the global variables in MDAppController.

I lessened the role that MDAppController plays in setting up the default values for the MDHLDocument-related defaults, like the view mode, whether to show inspector, or invisible items, etc. The registering of default values was moved from MDAppController's +initialize method to MDHLDocument's +initialize method.

Updated the MDViewOptionsController, MDInspectorController, and MDQuickLookController classes to work with the new functionality of MDHLDocument.