Source Finagler: BUGFIX: fixed bug that prevented remembering whether View Options, Quick Look, and Inspector should be shown from one launch to the next.
The MDViewOptionsController, MDInspectorController, and MDQuickLookController all listen for a notification of when their respective window closes, and when it does, they set the respective value in user defaults about whether they should be shown or not. However, if any of them were open when a user quits the app, those windows would be closed as part of the normal app shutdown process, and so the classes would always record that they should not be shown. These classes have been updated to now use an `appIsTerminating` BOOL flag to keep track of whether the app is about to quit (by listening for the NSApplicationWillTerminateNotification notification), and consults the value of that flag before blindly saving the in user defaults.